+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/source/preload/BasePreloadManager.java
@@ -33,10 +33,11 @@ import androidx.media3.common.util.Util;
 import androidx.media3.exoplayer.source.MediaSource;
 import com.google.common.base.Predicate;
 import com.google.common.base.Supplier;
+import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
-import java.util.PriorityQueue;
 
 /**
  * A base implementation of a preload manager, which maintains the lifecycle of {@linkplain
@@ -74,7 +75,13 @@ public abstract class BasePreloadManager<T, PreloadStatusT> {
   private final Handler applicationHandler;
 
   @GuardedBy("lock")
-  private final PriorityQueue<MediaSourceHolder> sourceHolderPriorityQueue;
+  private final List<MediaSourceHolder> sourceHolderPriorityList;
+
+  @GuardedBy("lock")
+  private int indexForSourceHolderToPreload;
+
+  @GuardedBy("lock")
+  private int indexForSourceHolderToClear;
 
   @GuardedBy("lock")
   @Nullable
@@ -95,8 +102,8 @@ public abstract class BasePreloadManager<T, PreloadStatusT> {
     listeners =
         new ListenerSet<>(applicationHandler.getLooper(), Clock.DEFAULT, (listener, flags) -> {});
     mediaSourceHolderMap = new MediaSourceHolderMap();
-    sourceHolderPriorityQueue = new PriorityQueue<>();
     this.rankingDataComparator.setInvalidationListener(this::invalidate);
+    sourceHolderPriorityList = new ArrayList<>();
   }
 
   /**
@@ -200,12 +207,21 @@ public abstract class BasePreloadManager<T, PreloadStatusT> {
    */
   public final void invalidate() {
     synchronized (lock) {
-      sourceHolderPriorityQueue.clear();
-      sourceHolderPriorityQueue.addAll(mediaSourceHolderMap.values());
-      while (!sourceHolderPriorityQueue.isEmpty() && !maybeStartPreloadingNextSourceHolder()) {
-        sourceHolderPriorityQueue.poll();
+      resetSourceHolderPriorityList();
+      while (indexForSourceHolderToPreload < sourceHolderPriorityList.size()
+          && !maybeStartPreloadingNextSourceHolder()) {
+        indexForSourceHolderToPreload++;
       }
     }
+  }
+
+  @GuardedBy("lock")
+  private void resetSourceHolderPriorityList() {
+    sourceHolderPriorityList.clear();
+    sourceHolderPriorityList.addAll(mediaSourceHolderMap.values());
+    Collections.sort(sourceHolderPriorityList);
+    indexForSourceHolderToPreload = 0;
+    indexForSourceHolderToClear = sourceHolderPriorityList.size() - 1;
   }
 
   /**
@@ -316,7 +332,7 @@ public abstract class BasePreloadManager<T, PreloadStatusT> {
     }
     mediaSourceHolderMap.clear();
     synchronized (lock) {
-      sourceHolderPriorityQueue.clear();
+      resetSourceHolderPriorityList();
       targetPreloadStatusOfCurrentPreloadingSource = null;
     }
   }
@@ -434,6 +450,19 @@ public abstract class BasePreloadManager<T, PreloadStatusT> {
         });
   }
 
+  /** Called when there is a {@link MediaSource} has been cleared. */
+  protected final void onSourceCleared() {
+    synchronized (lock) {
+      indexForSourceHolderToClear--;
+    }
+  }
+
+  /**
+   * Called when the given {@link MediaItem} has its corresponding {@link MediaSource} updated.
+   *
+   * @param mediaItem The {@link MediaItem} that apps have added with.
+   * @param updatedMediaSource The updated {@link MediaSource}.
+   */
   protected final void onMediaSourceUpdated(MediaItem mediaItem, MediaSource updatedMediaSource) {
     postOrRun(
         applicationHandler,
@@ -453,18 +482,29 @@ public abstract class BasePreloadManager<T, PreloadStatusT> {
   private void maybeAdvanceToNextMediaSourceHolder() {
     synchronized (lock) {
       do {
-        sourceHolderPriorityQueue.poll();
-      } while (!sourceHolderPriorityQueue.isEmpty() && !maybeStartPreloadingNextSourceHolder());
+        indexForSourceHolderToPreload++;
+      } while (indexForSourceHolderToPreload < sourceHolderPriorityList.size()
+          && !maybeStartPreloadingNextSourceHolder());
     }
   }
 
   @GuardedBy("lock")
   @Nullable
   private MediaSourceHolder getCurrentlyPreloadingMediaSourceHolder() {
-    if (sourceHolderPriorityQueue.isEmpty()) {
+    if (indexForSourceHolderToPreload >= sourceHolderPriorityList.size()) {
       return null;
     }
-    return sourceHolderPriorityQueue.peek();
+    return sourceHolderPriorityList.get(indexForSourceHolderToPreload);
+  }
+
+  @Nullable
+  protected MediaSourceHolder getMediaSourceHolderToClear() {
+    synchronized (lock) {
+      if (indexForSourceHolderToPreload >= indexForSourceHolderToClear) {
+        return null;
+      }
+      return sourceHolderPriorityList.get(indexForSourceHolderToClear);
+    }
   }
 
   @Nullable
@@ -537,16 +577,16 @@ public abstract class BasePreloadManager<T, PreloadStatusT> {
   protected void releaseInternal() {}
 
   /**
-   * Starts to preload the {@link MediaSource} at the head of the priority queue.
+   * Starts to preload the {@link MediaSource} at the head of the priority list.
    *
-   * @return {@code true} if the {@link MediaSource} at the head of the priority queue starts to
+   * @return {@code true} if the {@link MediaSource} at the head of the priority list starts to
    *     preload, otherwise {@code false}.
-   * @throws NullPointerException if the priority queue is empty.
    */
   @GuardedBy("lock")
   private boolean maybeStartPreloadingNextSourceHolder() {
     if (shouldStartPreloadingNextSource()) {
-      MediaSourceHolder preloadingHolder = checkNotNull(sourceHolderPriorityQueue.peek());
+      MediaSourceHolder preloadingHolder =
+          sourceHolderPriorityList.get(indexForSourceHolderToPreload);
       if (preloadingHolder.isReleased()) {
         return false;
       }